requirements.txt
# Core dependencies
perplexityai>=0.1.0
python-dotenv>=1.0.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Templating
jinja2>=3.1.0

# Data handling
pandas>=2.0.0
pydantic>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# CLI dependencies
prompt-toolkit>=3.0.0
rich>=13.0.0
click>=8.1.0

# Utilities
requests>=2.31.0
python-slugify>=8.0.0

# Packaging (optional)
pyinstaller>=6.0.0
